Renamed connection-menu-items to connection-menu-mobile
