+++ software/web_ui/src/lib/components/connection-menu-mobile.svelte
